refactor: migrate dashboard widgets and skeletons to packages
