--- view/packages/components/dashboard/system/skeletons/index.tsx
@@ -1,7 +1,0 @@
-export { CPUUsageCardSkeleton, CPUUsageCardSkeletonContent } from './cpu-usage';
-export { LoadAverageCardSkeleton, LoadAverageCardSkeletonContent } from './load-average';
-export { MemoryUsageCardSkeleton, MemoryUsageCardSkeletonContent } from './memory-usage';
-export { DiskUsageCardSkeleton, DiskUsageCardSkeletonContent } from './disk-usage';
-export { SystemInfoCardSkeleton } from './system-info';
-export { ClockCardSkeletonContent } from './clock';
-export { NetworkCardSkeletonContent } from './network';
