--- view/packages/components/dashboard/utils/truncate-id.ts
@@ -1,5 +1,0 @@
-const truncateId = (id: string) => {
-  return id?.substring(0, 12) || '';
-};
-
-export default truncateId;
