+++ view/packages/hooks/dashboard/use-container.tsx
@@ -3,8 +3,7 @@ import { ContainerData } from '@/redux/types/monitor';
 import { useTranslation } from '@/packages/hooks/shared/use-translation';
 import { TypographyMuted, TypographySmall } from '@/components/ui/typography';
 import { Badge } from '@/components/ui/badge';
-import truncateId from '@/packages/components/dashboard/utils/truncate-id';
-import getStatusColor from '@/packages/components/dashboard/utils/get-status-color';
+import { truncateId, getStatusColor } from '@/packages/utils/dashboard';
 
 export function useContainer() {
   const { t } = useTranslation();
