+++ view/packages/utils/dashboard.ts
@@ -1,6 +1,56 @@
-import { CHART_COLORS } from './constants';
 import { BarChartDataItem } from '@/components/ui/bar-chart-component';
 import { DoughnutChartDataItem } from '@/components/ui/doughnut-chart-component';
+
+// Color constants for charts and visualizations
+export const CHART_COLORS = {
+  blue: '#3b82f6',
+  green: '#10b981',
+  orange: '#f59e0b',
+  red: '#ef4444',
+  purple: '#a855f7',
+  yellow: '#eab308'
+};
+
+// Default values for system metrics
+export const DEFAULT_METRICS = {
+  load: {
+    oneMin: 0 as number,
+    fiveMin: 0 as number,
+    fifteenMin: 0 as number
+  },
+  cpu: {
+    overall: 0 as number,
+    per_core: [] as Array<{
+      core_id: number;
+      usage: number;
+    }>
+  },
+  memory: {
+    total: 0 as number,
+    used: 0 as number,
+    percentage: 0 as number
+  },
+  disk: {
+    percentage: 0 as number,
+    used: 0 as number,
+    total: 0 as number,
+    allMounts: [] as any[]
+  }
+};
+
+export const getStatusColor = (status: string) => {
+  if (status?.toLowerCase().includes('running')) return 'bg-green-100 text-green-800 rounded-full';
+
+  if (status?.toLowerCase().includes('exited')) return 'bg-red-100 text-red-800 rounded-full';
+
+  if (status?.toLowerCase().includes('created')) return 'bg-blue-100 text-blue-800 rounded-full';
+
+  return 'bg-gray-100 text-gray-800 rounded-full';
+};
+
+export const truncateId = (id: string) => {
+  return id?.substring(0, 12) || '';
+};
 
 export const formatGB = (value: number): string => `${value.toFixed(2)}`;
 export const formatPercentage = (value: number): string => `${value.toFixed(1)}`;
@@ -67,7 +117,10 @@ export const createMemoryChartConfig = () => ({
 });
 
 export const createCPUChartData = (
-  perCore: Array<{ core_id: number; usage: number }>
+  perCore: Array<{
+    core_id: number;
+    usage: number;
+  }>
 ): BarChartDataItem[] => {
   if (!perCore || perCore.length === 0) {
     return [];
@@ -90,7 +143,13 @@ export const createCPUChartData = (
 };
 
 export const createCPUChartConfig = (coreCount: number) => {
-  const config: Record<string, { label: string; color: string }> = {};
+  const config: Record<
+    string,
+    {
+      label: string;
+      color: string;
+    }
+  > = {};
   const colors = [
     CHART_COLORS.blue,
     CHART_COLORS.green,
